move add phone at setup spec to its own file
